highly experimental stuff

+++ math/vec2declare.h
@@ -2,5 +2,5 @@
 template <class type = float> struct vec2t;
 
 typedef vec2t<int> vec2i;
-typedef vec2t<float> vec2;
+typedef vec2t<double> vec2;
 typedef vec2t<double> vec2d;
